+++ ChangeLog.md
@@ -1,3 +1,7 @@
+###### v0.8.9 (2020-08-30):
+ * Added summary table display.
+ * Updated parsing function to convert to percent.
+
 ###### v0.8.8 (2020-08-30):
  * Added function for calculating summary table data to the IAOReader class.
  * Added Shiny handler for updating the plot.